Move serialization code out of AllocatedHosts
