+++ config-provisioning/src/main/java/com/yahoo/config/provision/serialization/package-info.java
@@ -1,0 +1,5 @@
+// Copyright 2019 Oath Inc. Licensed under the terms of the Apache 2.0 license. See LICENSE in the project root.
+@ExportPackage
+package com.yahoo.config.provision.serialization;
+
+import com.yahoo.osgi.annotation.ExportPackage;
